requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.1
Flask-JWT-Extended>=4.6
Flask-Migrate>=4.0

redis>=5.0
stripe>=8.0

gunicorn>=21.2

Flask-Cors>=4.0
Flask-Mail>=0.9.1

python-dotenv>=1.0

Werkzeug>=2.3

Flask-Limiter>=3.5
Flask-Limiter[redis]

python-json-logger>=2.0

celery==5.3.6
redis==5.0.1
flask-mail==0.9.1
